+++ Fabric/src/main/java/net/merchantpug/bovinesandbuttercups/platform/FabricPlatformHelper.java
@@ -1,6 +1,9 @@
 package net.merchantpug.bovinesandbuttercups.platform;
 
 import com.mojang.serialization.Codec;
+import net.merchantpug.bovinesandbuttercups.api.condition.ConditionType;
+import net.merchantpug.bovinesandbuttercups.api.condition.block.BlockConditionType;
+import net.merchantpug.bovinesandbuttercups.api.condition.entity.EntityConditionType;
 import net.merchantpug.bovinesandbuttercups.api.type.CowType;
 import net.merchantpug.bovinesandbuttercups.data.ConfiguredCowTypeRegistry;
 import net.merchantpug.bovinesandbuttercups.platform.services.IPlatformHelper;
@@ -43,6 +46,16 @@ public class FabricPlatformHelper implements IPlatformHelper {
     @Override
     public Codec<CowType<?>> getCowTypeCodec() {
         return BovineRegistriesFabric.COW_TYPE.byNameCodec();
+    }
+
+    @Override
+    public Codec<EntityConditionType<?>> getEntityConditionTypeCodec() {
+        return BovineRegistriesFabric.ENTITY_CONDITION_TYPE.byNameCodec();
+    }
+
+    @Override
+    public Codec<BlockConditionType<?>> getBlockConditionTypeCodec() {
+        return BovineRegistriesFabric.BLOCK_CONDITION_TYPE.byNameCodec();
     }
 
     @Override